+++ plugins/source/gitlab/resources/services/projects/project_members.go
@@ -14,14 +14,7 @@ func ProjectMembers() *schema.Table {
 		Name:      "gitlab_project_members",
 		Resolver:  fetchProjectMembers,
 		Transform: client.TransformWithStruct(&gitlab.ProjectMember{}, transformers.WithPrimaryKeys("ID")),
-		Columns: schema.ColumnList{client.BaseURLColumn,
-			{
-				Name:            "project_id",
-				Type:            schema.TypeInt,
-				Resolver:        resolveProjectID,
-				CreationOptions: schema.ColumnCreationOptions{NotNull: true, PrimaryKey: true},
-			},
-		},
+		Columns:   schema.ColumnList{client.BaseURLColumn, projectIDColumn},
 	}
 }
