--- plugins/source/gitlab/resources/services/projects/projects_fetch.go
@@ -1,38 +1,0 @@
-package projects
-
-import (
-	"context"
-
-	"github.com/cloudquery/cloudquery/plugins/source/gitlab/client"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
-	"github.com/xanzy/go-gitlab"
-)
-
-func fetchProjects(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	c := meta.(*client.Client)
-
-	opt := &gitlab.ListProjectsOptions{
-		MinAccessLevel: c.MinAccessLevel,
-		ListOptions: gitlab.ListOptions{
-			PerPage: 1000,
-		},
-	}
-	for {
-		// Get the first page with projects.
-		projects, resp, err := c.Gitlab.Projects.ListProjects(opt, gitlab.WithContext(ctx))
-		if err != nil {
-			return err
-		}
-		res <- projects
-
-		// Exit the loop when we've seen all pages.
-		if resp.NextPage == 0 {
-			break
-		}
-
-		// Update the page number to get the next page.
-		opt.Page = resp.NextPage
-	}
-
-	return nil
-}
