+++ plugins/source/gitlab/resources/services/projects/projects_releases.go
@@ -1,6 +1,8 @@
 package projects
 
 import (
+	"context"
+
 	"github.com/cloudquery/cloudquery/plugins/source/gitlab/client"
 	"github.com/cloudquery/plugin-sdk/v2/schema"
 	"github.com/cloudquery/plugin-sdk/v2/transformers"
@@ -12,13 +14,34 @@ func ProjectsReleases() *schema.Table {
 		Name:      "gitlab_projects_releases",
 		Resolver:  fetchProjectsReleases,
 		Transform: client.TransformWithStruct(&gitlab.Release{}, transformers.WithPrimaryKeys("CreatedAt")),
-		Columns: schema.ColumnList{client.BaseURLColumn,
-			{
-				Name:            "project_id",
-				Type:            schema.TypeInt,
-				Resolver:        resolveProjectID,
-				CreationOptions: schema.ColumnCreationOptions{NotNull: true, PrimaryKey: true},
-			},
+		Columns:   schema.ColumnList{client.BaseURLColumn, projectIDColumn},
+	}
+}
+
+func fetchProjectsReleases(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	c := meta.(*client.Client)
+	project := parent.Item.(*gitlab.Project)
+	opt := &gitlab.ListReleasesOptions{
+		ListOptions: gitlab.ListOptions{
+			PerPage: 1000,
 		},
 	}
+
+	for {
+		// Get the first page with projects.
+		members, resp, err := c.Gitlab.Releases.ListReleases(project.ID, opt, gitlab.WithContext(ctx))
+		if err != nil {
+			return err
+		}
+		res <- members
+		// Exit the loop when we've seen all pages.
+		if resp.NextPage == 0 {
+			break
+		}
+
+		// Update the page number to get the next page.
+		opt.Page = resp.NextPage
+	}
+
+	return nil
 }
