--- plugins/source/gitlab/resources/services/settings/settings_fetch.go
@@ -1,25 +1,0 @@
-package settings
-
-import (
-	"context"
-
-	"github.com/cloudquery/cloudquery/plugins/source/gitlab/client"
-	"github.com/cloudquery/plugin-sdk/v2/schema"
-	"github.com/xanzy/go-gitlab"
-)
-
-func fetchSettings(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	c := meta.(*client.Client)
-	if c.BaseURL == "" {
-		c.Logger().Info().Str("table", "gitlab_settings").Msg("not supported for GitLab SaaS, skipping...")
-		return nil
-	}
-	setting, _, err := c.Gitlab.Settings.GetSettings(gitlab.WithContext(ctx))
-	if err != nil {
-		return err
-	}
-
-	res <- setting
-
-	return nil
-}
